+++ src/MainLib/Marqdouj.DotNet.Web.JsInterop.Azure.Maps/Interop/Modules/AzAnimations.cs
@@ -1,4 +1,5 @@
 ﻿using Marqdouj.DotNet.Web.JsInterop.Azure.Maps.Models.Animations;
+using Marqdouj.DotNet.Web.JsInterop.GeoJson;
 using Microsoft.JSInterop;
 using System.Runtime.CompilerServices;
 
@@ -18,6 +19,17 @@ namespace Marqdouj.DotNet.Web.JsInterop.Azure.Maps.Interop.Modules
         ValueTask AnimateShape(string mapId, ShapeAnimationOptions options);
 
         /// <summary>
+        /// Extracts points from a shape or feature that form a time based route, and sorts them by time.
+        /// Timestamps must parsable by the `atlas.math.parseTimestamp` function.
+        /// Features must be a Point, MultiPoint, or LineString and must contain properties that include timestamp information.
+        /// </summary>
+        /// <param name="mapId">The map id.</param>
+        /// <param name="shapeId">The shape id which contains the time based route.</param>
+        /// <param name="length">The number of elements to return. -1 returns all elements.</param>
+        /// <param name="timestampProperty">the property name that contains timestamp information.  If not specified, `_timestamp` will be used.</param>
+        ValueTask<List<Feature<Point, object?>[]>?> ExtractRoutePoints(string mapId, string shapeId, int length, string? timestampProperty);
+
+        /// <summary>
         /// Retrieves the name of all the built in easing functions.
         /// </summary>
         /// <param name="mapId"></param>
@@ -29,16 +41,22 @@ namespace Marqdouj.DotNet.Web.JsInterop.Azure.Maps.Interop.Modules
     {
         private readonly Lazy<Task<IJSObjectReference>> moduleTask = moduleTask;
 
-        public async ValueTask<List<string>> GetEasingNames(string mapId)
-        {
-            var module = await moduleTask.Value;
-            return await module.InvokeAsync<List<string>>(GetJsInteropMethod(), mapId);
-        }
-
         public async ValueTask AnimateShape(string mapId, ShapeAnimationOptions options)
         {
             var module = await moduleTask.Value;
             await module.InvokeVoidAsync(GetJsInteropMethod(), mapId, options);
+        }
+
+        public async ValueTask<List<Feature<Point, object?>[]>?> ExtractRoutePoints(string mapId, string shapeId, int length, string? timestampProperty)
+        {
+            var module = await moduleTask.Value;
+            return await module.InvokeAsync<List<Feature<Point, object?>[]>?>(GetJsInteropMethod(), mapId, shapeId, length, timestampProperty);
+        }
+
+        public async ValueTask<List<string>> GetEasingNames(string mapId)
+        {
+            var module = await moduleTask.Value;
+            return await module.InvokeAsync<List<string>>(GetJsInteropMethod(), mapId);
         }
 
         private static string GetJsInteropMethod([CallerMemberName] string name = "")