+++ src/MainLib/Sandbox/Components/Pages/AzureMaps/Common/MapProviderHelper.cs
@@ -1,4 +1,5 @@
-﻿using Marqdouj.DotNet.Web.JsInterop.Azure.Maps.Interop;
+﻿using Marqdouj.DotNet.Web.JsInterop.Azure.Maps.Components;
+using Marqdouj.DotNet.Web.JsInterop.Azure.Maps.Interop;
 using Marqdouj.DotNet.Web.JsInterop.Azure.Maps.Models.Events;
 using Microsoft.FluentUI.AspNetCore.Components;
 using Sandbox.UI.Models;